requirements.txt
flask~=3.0.3
flask-cors~=3.0.10
werkzeug~=3.0.3
sqlalchemy~=2.0.30
tensorflow~=2.14.0
numpy~=1.26.4
python-dotenv~=1.0.1
psycopg2-binary